Refactored the CPU out of the main
